Refactor positivity data simulator location
