+++ causallib/positivity/tests/test_base.py
@@ -1,6 +1,5 @@
 import unittest
-from causallib.positivity import base_positivity
-from causallib.positivity.tests.testdata import make_1d_overlap_data
+from causallib.positivity.datasets.test_data_simulator import make_1d_overlap_data
 
 
 class TestBase(unittest.TestCase):
